Refactor common code into IdentityProviderDialog
