+++ src/dialogs/AzureADDialog.ts
@@ -22,133 +22,38 @@
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 import * as builder from "botbuilder";
-import * as dialogs from "botbuilder-dialogs";
-import { IOAuth2Provider } from "../providers";
+import { IOAuth2Provider, AzureADv1Provider } from "../providers";
+import { IdentityProviderDialog } from "./IdentityProviderDialog";
 
 export const AZUREAD_DIALOG = "AzureADDialog";
-const CHOICE_PROMPT = "ChoicePrompt";
-const OAUTH_PROMPT = "OAuthPrompt";
-const MAIN_WATERFALL_DIALOG = "MainWaterfallDialog";
-const SHOWPROFILE_FLOW = "ShowProfileFlow";
-const SIGNIN_FLOW = "SigninFlow";
-const SIGNOUT_FLOW = "SignoutFlow";
-const MAINLOOP_FLOW = "MainLoopFlow";
 
-export class AzureADDialog extends dialogs.ComponentDialog {
+export class AzureADDialog extends IdentityProviderDialog {
+
+    private provider: IOAuth2Provider;
 
     constructor(
-        private connectionName: string,
-        private provider: IOAuth2Provider)
+        connectionName: string)
     {
-        super(AZUREAD_DIALOG);
+        super(AZUREAD_DIALOG, connectionName);
 
-        this.addDialog(new dialogs.OAuthPrompt(OAUTH_PROMPT, {
-            connectionName,
-            text: "Please sign in so I can show you your profile.",
-            title: "Sign in",
-            timeout: 300000
-        }));
-        this.addDialog(new dialogs.ChoicePrompt(CHOICE_PROMPT));
-        this.addDialog(new dialogs.WaterfallDialog(MAIN_WATERFALL_DIALOG, [
-            this.getTokenStep.bind(this),
-            this.showProfileStep.bind(this),
-            this.restartMainLoopStep.bind(this),
-        ]));
-        this.addDialog(new dialogs.WaterfallDialog(MAINLOOP_FLOW, [
-            this.chooseActionStep.bind(this),
-            this.processActionStep.bind(this),
-            this.restartMainLoopStep.bind(this),
-        ]));
-        this.addDialog(new dialogs.WaterfallDialog(SHOWPROFILE_FLOW, [
-            this.getTokenStep.bind(this),
-            this.showProfileStep.bind(this),
-        ]));
-        this.addDialog(new dialogs.WaterfallDialog(SIGNOUT_FLOW, [
-            this.signOutStep.bind(this),
-        ]));
-        this.addDialog(new dialogs.WaterfallDialog(SIGNIN_FLOW, [
-            this.signOutStep.bind(this),
-            this.getTokenStep.bind(this),
-            this.showProfileStep.bind(this),
-        ]));
-
-        this.initialDialogId = MAIN_WATERFALL_DIALOG;
+        this.provider = new AzureADv1Provider(null, null);
     }
 
-    private getTokenStep(stepContext: dialogs.WaterfallStepContext) {
-        return stepContext.beginDialog(OAUTH_PROMPT);
+    public get displayName() { return "Azure AD"; }
+
+    public async getProfileAsync(accessToken: string): Promise<any> {
+        return await this.provider.getProfileAsync(accessToken);
     }
 
-    private async showProfileStep(stepContext: dialogs.WaterfallStepContext) {
-        const tokenResponse: builder.TokenResponse = stepContext.result;
-        if (tokenResponse) {
-            const profile = await this.provider.getProfileAsync(tokenResponse.token);
-            const card = builder.CardFactory.thumbnailCard(
-                profile.displayName,
-                `<b>E-mail</b>: ${profile.mail}<br/>
-                <b>Title</b>: ${profile.jobTitle}<br/>
-                <b>Office location</b>: ${profile.officeLocation}`,
-                [],
-                [],
-                { subtitle: profile.userPrincipalName });
-            await stepContext.context.sendActivity({ 
-                text: `Here's your profile in ${this.provider.displayName}`,
-                attachments: [ card ],
-            });
-            return await stepContext.next();
-        } else {
-            await stepContext.context.sendActivity("Please sign in so I can show you your profile.");
-            return await stepContext.replaceDialog(MAIN_WATERFALL_DIALOG);
-        }
-    }
-
-    private async chooseActionStep(stepContext: dialogs.WaterfallStepContext) {
-        const adapter = stepContext.context.adapter as builder.BotFrameworkAdapter;
-        const tokenResponse = await adapter.getUserToken(stepContext.context, this.connectionName);
-
-        let choices = [];
-        if (tokenResponse && tokenResponse.token) {
-            choices = [ "Show profile", "Sign out", "Back" ];
-        } else {
-            choices = [ "Sign in", "Back" ];
-        }
-
-        return await stepContext.prompt(CHOICE_PROMPT, {
-            choices: dialogs.ChoiceFactory.toChoices(choices),
-        });
-    }
-
-    private async processActionStep(stepContext: dialogs.WaterfallStepContext) {
-        const choice = stepContext.result.value;
-        switch (choice) {
-            case "Show profile":
-                return await stepContext.beginDialog(SHOWPROFILE_FLOW);
-
-            case "Sign out":
-                return await stepContext.beginDialog(SIGNOUT_FLOW);
-
-            case "Sign in":
-                return await stepContext.beginDialog(SIGNIN_FLOW, { silent: true });
-
-            case "Back":
-                return await stepContext.endDialog();
-
-            default:
-                await stepContext.context.sendActivity(`"I didn't recognize your choice '${choice}'.`);
-                return await stepContext.next();
-        }
-    }
-
-    private async signOutStep(stepContext: dialogs.WaterfallStepContext<any>) {
-        const adapter = stepContext.context.adapter as builder.BotFrameworkAdapter;
-        await adapter.signOutUser(stepContext.context, this.connectionName);
-        if (!stepContext.options.silent) {
-            await stepContext.context.sendActivity(`You're now signed out of ${this.provider.displayName}.`);
-        }
-        return await stepContext.next();
-    }
-
-    private restartMainLoopStep(stepContext: dialogs.WaterfallStepContext) {
-        return stepContext.replaceDialog(MAINLOOP_FLOW);
+    protected async getProfileCardAsync(accessToken: string): Promise<builder.Attachment> {
+        const profile = await this.getProfileAsync(accessToken);
+        return builder.CardFactory.thumbnailCard(
+            profile.displayName,
+            `<b>E-mail</b>: ${profile.mail}<br/>
+            <b>Title</b>: ${profile.jobTitle}<br/>
+            <b>Office location</b>: ${profile.officeLocation}`,
+            [],
+            [],
+            { subtitle: profile.userPrincipalName });
     }
 }
